Wip/status service

+++ src/tateyama/oltp/control.cpp
@@ -32,6 +32,7 @@
 #include "configuration.h"
 #include "proc_mutex.h"
 #include "monitor.h"
+#include "status_info.h"
 
 DECLARE_string(conf);
 DEFINE_bool(quiesce, false, "invoke in quiesce mode");  // NOLINT for quiesce
@@ -43,25 +44,34 @@ namespace tateyama::bootstrap {
 
 constexpr std::string_view server_name_string = "tateyama-server";
 constexpr std::string_view server_name_string_for_status = "Tsurugi OLTP database";
-const std::size_t shutdown_check_count = 50;
+const std::size_t check_count = 100;
+const int sleep_time_unit = 20;
 
 using namespace tateyama::bootstrap::utils;
 
 static bool status_check(proc_mutex::lock_state state, const boost::filesystem::path& lock_file) {
-    auto file_mutex = std::make_unique<proc_mutex>(lock_file, false);
-    for (size_t i = 0; i < shutdown_check_count; i++) {
-        usleep(100 * 1000);
+    std::unique_ptr<proc_mutex> file_mutex{};
+    for (size_t i = 0; i < check_count; i++) {
+        if (!file_mutex) {
+            try {
+                file_mutex = std::make_unique<proc_mutex>(lock_file, false);
+            } catch (std::runtime_error &e) {
+                usleep(sleep_time_unit * 1000);
+                continue;
+            }
+        }
         if (file_mutex->check() == state) {
             return true;
         }
+        usleep(sleep_time_unit * 5 * 1000);
     }
     return false;
 }
 
-
 int oltp_start([[maybe_unused]] int argc, char* argv[], char *argv0, bool need_check) {
     std::string server_name(server_name_string);
-    if (auto pid = fork(); pid == 0) {
+    pid_t child_pid = 0;
+    if (child_pid = fork(); child_pid == 0) {
         boost::filesystem::path path_for_this{};
         if (auto a0f = boost::filesystem::path(argv0); a0f.parent_path().string().empty()) {
             path_for_this = boost::filesystem::canonical(boost::process::search_path(a0f));
@@ -81,32 +91,104 @@ int oltp_start([[maybe_unused]] int argc, char* argv[], char *argv0, bool need_c
         LOG(ERROR) << "cannot find the location of " << argv0;
         exit(1);
     } else {  // to output pid to DVLOG(log_trace)
-        DVLOG(log_trace) << "start " << server_name << ", pid = " << pid;
+        DVLOG(log_trace) << "start " << server_name << ", pid = " << child_pid;
     }
 
+    int rc = tateyama::bootstrap::return_code::ok;
     if (need_check) {
+        std::unique_ptr<utils::monitor> monitor_output{};
+
         // command arguments
         gflags::SetUsageMessage("tateyama database server CLI");
         gflags::ParseCommandLineFlags(&argc, &argv, true);
 
+        if(!FLAGS_monitor.empty()) {
+            monitor_output = std::make_unique<utils::monitor>(FLAGS_monitor);
+            monitor_output->start();
+        }
+
         auto bst_conf = utils::bootstrap_configuration(FLAGS_conf);
         if (auto conf = bst_conf.create_configuration(); conf != nullptr) {
-            usleep(100 * 1000);
-            try {
-                if (status_check(proc_mutex::lock_state::locked, bst_conf.lock_file())) {
-                    return 0;
+            std::size_t n = 0;
+
+            std::unique_ptr<proc_mutex> file_mutex{};
+            enum {
+                init,
+                ok,
+                another,
+            } check_result = init;
+            for (size_t i = n; i < check_count; i++) {
+                if (!file_mutex) {
+                    try {
+                        file_mutex = std::make_unique<proc_mutex>(bst_conf.lock_file(), false);
+                    } catch (std::runtime_error &e) {
+                        usleep(sleep_time_unit * 1000);
+                        continue;
+                    }
                 }
-                LOG(ERROR) << "cannot invoke a server process";
-                return 1;
-            } catch (std::runtime_error &e) {
-                LOG(ERROR) << e.what();
-                return 2;
+                if (file_mutex->check() == proc_mutex::lock_state::locked) {
+                    auto pid = file_mutex->pid(); 
+                    if (pid == 0) {
+                        continue;
+                    }
+                    check_result = (child_pid == pid) ? ok : another;
+                    n = i;
+                    break;
+                }
+                usleep(sleep_time_unit * 5 * 1000);
             }
+
+            if (check_result == ok) {
+                std::unique_ptr<status_info_brigde> status_info = std::make_unique<status_info_brigde>();
+                // wait for creation of shared memory for status info
+                for (size_t i = n; i < check_count; i++) {
+                    if (status_info->attach(bst_conf.digest())) {
+                        n = i;
+                        break;
+                    }
+                    usleep(sleep_time_unit * 1000);
+                }
+
+                // wait for pid set
+                bool checked = false;
+                for (size_t i = n; i < check_count; i++) {
+                    auto pid = status_info->pid();
+                    if (pid == 0) {
+                        usleep(sleep_time_unit * 1000);
+                        continue;
+                    }
+                    if (child_pid == pid) {
+                        if (monitor_output) {
+                            monitor_output->finish(true);
+                        }
+                        return tateyama::bootstrap::return_code::ok;
+                    }
+                    LOG(ERROR) << "another " << server_name_string_for_status << " is running";
+                    rc = tateyama::bootstrap::return_code::err;
+                    checked = true;
+                }
+                if (!checked) {
+                    LOG(ERROR) << "cannot confirm the server process within the specified time";
+                    rc = tateyama::bootstrap::return_code::err;
+                }
+            } else {
+                if (check_result == init) {
+                    LOG(ERROR) << "cannot invoke a server process";
+                } else {
+                    LOG(ERROR) << "another " << server_name_string_for_status << " is running";
+                }
+                rc = tateyama::bootstrap::return_code::err;
+            }
+        } else {
+            LOG(ERROR) << "cannot find the configuration file";
+            rc = tateyama::bootstrap::return_code::err;
         }
-        LOG(ERROR) << "error in create_configuration";
-        return 3;
+        if (monitor_output) {
+            monitor_output->finish(false);
+        }
+        return rc;
     }
-    return 0;
+    return tateyama::bootstrap::return_code::ok;  // when need_check is false, it should return ok
 }
 
 int oltp_shutdown_kill(int argc, char* argv[], bool force, bool status_output) {
@@ -132,7 +214,7 @@ int oltp_shutdown_kill(int argc, char* argv[], bool force, bool status_output) {
                 if (pid != 0) {
                     DVLOG(log_trace) << "kill (SIGKILL) to process " << pid << " and remove " << file_mutex->name();
                     kill(pid, SIGKILL);
-                    usleep(100 * 1000);
+                    usleep(sleep_time_unit * 5 * 1000);
                     if (monitor_output) {
                         monitor_output->finish(true);
                     }
@@ -141,20 +223,26 @@ int oltp_shutdown_kill(int argc, char* argv[], bool force, bool status_output) {
                 LOG(ERROR) << "contents of the file (" << file_mutex->name() << ") cannot be used";
                 rc = tateyama::bootstrap::return_code::err;
             } else {
-                auto pid = file_mutex->pid(true);
-                if (pid != 0) {
-                    DVLOG(log_trace) << "kill (SIGINT) to process " << pid;
-                    kill(pid, SIGINT);
-                    if (status_check(proc_mutex::lock_state::no_file, bst_conf.lock_file())) {
-                        if (monitor_output) {
-                            monitor_output->finish(true);
+                std::unique_ptr<status_info_brigde> status_info = std::make_unique<status_info_brigde>(bst_conf.digest());
+                if (!status_info->shutdown()) {
+                    auto pid = file_mutex->pid(true);
+                    if (pid != 0) {
+                        DVLOG(log_trace) << "kill (SIGINT) to process " << pid;
+                        kill(pid, SIGINT);
+                        if (status_check(proc_mutex::lock_state::no_file, bst_conf.lock_file())) {
+                            if (monitor_output) {
+                                monitor_output->finish(true);
+                            }
+                            return rc;
                         }
-                        return rc;
+                        LOG(ERROR) << "failed to shutdown, the server may still be alive";
+                        rc = tateyama::bootstrap::return_code::err;
+                    } else {
+                        LOG(ERROR) << "contents of the file (" << file_mutex->name() << ") cannot be used";
+                        rc = tateyama::bootstrap::return_code::err;
                     }
-                    LOG(ERROR) << "failed to shutdown, the server may still be alive";
-                    rc = tateyama::bootstrap::return_code::err;
                 } else {
-                    LOG(ERROR) << "contents of the file (" << file_mutex->name() << ") cannot be used";
+                    LOG(ERROR) << "another shutdown is being conducted";
                     rc = tateyama::bootstrap::return_code::err;
                 }
             }
@@ -196,42 +284,58 @@ int oltp_status(int argc, char* argv[]) {
                 monitor_output->status(tateyama::bootstrap::utils::status::stop);
                 break;
             }
-            std::cout << server_name_string_for_status << " is inactive" << std::endl;
+            std::cout << server_name_string_for_status << " is INACTIVE" << std::endl;
             break;
         case state::locked:
-            if (monitor_output) {
-                monitor_output->status(tateyama::bootstrap::utils::status::running);
-                break;
-            }
-            std::cout << server_name_string_for_status << " is running" << std::endl;
-            break;
-        case state::not_locked:  // this is an error that makes it difficult to understand the situation.
         {
-            if (monitor_output) {
-                monitor_output->status(tateyama::bootstrap::utils::status::disconnected);
-                break;
-            }
-            std::stringstream ss;
-            FILE *fp;
-            ss << "ps -ef | grep " << server_name_string << " | grep " << file_mutex->pid(false) << " | grep -v grep | wc -l";
-            if(fp = popen(ss.str().c_str(), "r"); fp == nullptr){
-                std::cout << "the service is unknown" << std::endl;
+            std::unique_ptr<status_info_brigde> status_info{};
+            try {
+                status_info = std::make_unique<status_info_brigde>(bst_conf.digest());
+
+                switch(status_info->whole()) {
+                case tateyama::status_info::state::initial:
+                case tateyama::status_info::state::ready:
+                    if (monitor_output) {
+                        monitor_output->status(tateyama::bootstrap::utils::status::ready);
+                        break;
+                    }
+                    std::cout << server_name_string_for_status << " is BOOTING_UP" << std::endl;
+                    break;
+                case tateyama::status_info::state::activated:
+                    if (monitor_output) {
+                        monitor_output->status(tateyama::bootstrap::utils::status::activated);
+                        break;
+                    }
+                    std::cout << server_name_string_for_status << " is RUNNING" << std::endl;
+                    break;
+                case tateyama::status_info::state::deactivating:
+                    if (monitor_output) {
+                        monitor_output->status(tateyama::bootstrap::utils::status::deactivating);
+                        break;
+                    }
+                    std::cout << server_name_string_for_status << " is  SHUTTING_DOWN" << std::endl;
+                    break;
+                case tateyama::status_info::state::deactivated:
+                    if (monitor_output) {
+                        monitor_output->status(tateyama::bootstrap::utils::status::deactivated);
+                        break;
+                    }
+                    std::cout << server_name_string_for_status << " is INACTIVE" << std::endl;
+                    break;
+                };
+            } catch (std::exception& e) {
+                LOG(ERROR) << e.what();
                 rc = tateyama::bootstrap::return_code::err;
-            } else {
-                int l;
-                if (fscanf(fp, "%d", &l) != 1) {
-                    std::cout << "the service is unknown" << std::endl;
-                    rc = tateyama::bootstrap::return_code::err;
-                } else {
-                    std::cout << server_name_string_for_status <<
-                        ((l == 0) ? " is booting up" : " is shutting down") << std::endl;
-                }
             }
             break;
         }
         default:
-            LOG(ERROR) << "error in proc_mutex check";
-            rc = tateyama::bootstrap::return_code::err;
+            if (monitor_output) {
+                monitor_output->status(tateyama::bootstrap::utils::status::unknown);
+                break;
+            }
+            std::cout << server_name_string_for_status << " is UNKNOWN" << std::endl;
+            break;
         }
         if (rc == tateyama::bootstrap::return_code::ok) {
             if (monitor_output) {
